Refatorando as classes e adicionada a classe Tarefas(corrigida) novamente
