--- FTP/teste.txt.txt
@@ -1,4 +1,0 @@
-dasdasdasdasdasd
-
-
-scuk lul
