+++ src/br/com/unisinos/redes/ftp/cliente/EnviadorDeComando.java
@@ -5,6 +5,15 @@ import java.io.PrintStream;
 import java.net.Socket;
 import java.util.Scanner;
 
+/**
+ * Classe responsável pelo envio de comandos para o servidor
+ * 
+ * @author 
+ * Gabriel Sperb Stoffel
+ * <br>
+ * Victor E. Scherer
+ *
+ */
 public class EnviadorDeComando implements Runnable {
 
 	private Socket socket;
@@ -17,26 +26,13 @@ public class EnviadorDeComando implements Runnable {
 	public void run() {
 
 		// Enviando dados para o servidor
-		try (PrintStream saida = new PrintStream(socket.getOutputStream())) {
+		try (PrintStream saida = new PrintStream(socket.getOutputStream()); Scanner scanner = new Scanner(System.in);) {
 
-			Scanner scanner = new Scanner(System.in);
-			
 			while (scanner.hasNextLine()) {
 				String comando = scanner.nextLine();
-
+				
 				if (comando.trim().toLowerCase().equals("help")) {
-					System.out.println(
-						"Comandos disponíveis:\n"
-						+ "put [/arquivo a enviar.txt] : Enviar arquivo\n"
-						+ "get [/arquivo a receber.txt] [/path receptor] : Receber arquivo\n" 
-						+ "rename [Nome antigo.txt] [Novo Nome] : Renomear arquivo\n"
-						+ "delete [Arquivo.txt] : Remover arquivo ou diretório\n" 
-						+ "cd [/caminho] : Mover-se por diretórios\n"
-						+ "ls : Exibir arquivos do Path atual\n" 
-						+ "pwd : Exibir diretório atual\n"
-						+ "mkdir [Diretório] : Criar diretório\n" 
-						+ "bye: Sair do FTP"
-					);
+					exibirComandos();
 				} else if (comando.trim().toLowerCase().equals("bye")) {
 					System.out.println("Até logo!");
 					break;
@@ -49,5 +45,20 @@ public class EnviadorDeComando implements Runnable {
 			e.printStackTrace();
 		}
 
+	}
+
+	private void exibirComandos() {
+		System.out.println(
+			"Comandos disponíveis:\n"
+			+ "put [/arquivo a enviar.txt] : Enviar arquivo\n"
+			+ "get [/arquivo a receber.txt] [/path receptor] : Receber arquivo\n" 
+			+ "rename [Nome antigo.txt] [Novo Nome] : Renomear arquivo\n"
+			+ "delete [Arquivo.txt] : Remover arquivo ou diretório\n" 
+			+ "cd [/caminho] : Mover-se por diretórios\n"
+			+ "ls : Exibir arquivos do Path atual\n" 
+			+ "pwd : Exibir diretório atual\n"
+			+ "mkdir [Diretório] : Criar diretório\n" 
+			+ "bye: Sair do FTP"
+		);
 	}
 }
