+++ src/br/com/unisinos/redes/ftp/servidor/Servidor.java
@@ -12,7 +12,10 @@ import br.com.unisinos.redes.ftp.tarefas.DistribuidorDeTarefas;
  * Classe que desempenhará o papel de Servidor onde abrirá a para os Clientes e
  * executará os comandos respectivos ao seu papel.
  * 
- * @author Gabriel Sperb Stoffel
+ * @author 
+ * Gabriel Sperb Stoffel
+ * <br>
+ * Victor E. Scherer
  *
  */
 public class Servidor {
