+++ src/br/com/unisinos/redes/ftp/tarefas/DistribuidorDeTarefas.java
@@ -4,6 +4,7 @@ import java.io.BufferedInputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
+import java.io.OutputStream;
 import java.io.PrintStream;
 import java.util.Scanner;
 
@@ -13,7 +14,10 @@ import br.com.unisinos.redes.ftp.cliente.Cliente;
  * Classe que desempenhará a função de capturar e distribuir os comando/tarefas
  * que serão executadas no Cliente ou Servidor.
  * 
- * @author Gabriel Sperb Stoffel
+ * @author 
+ * Gabriel Sperb Stoffel
+ * <br>
+ * Victor E. Scherer
  *
  */
 public class DistribuidorDeTarefas implements Runnable {
@@ -28,12 +32,11 @@ public class DistribuidorDeTarefas implements Runnable {
 	public void run() {
 
 		try (Scanner entradaCliente = new Scanner(cliente.getSocket().getInputStream());
-				PrintStream saidaCliente = new PrintStream(cliente.getSocket().getOutputStream())) {
+				PrintStream saidaCliente = new PrintStream(cliente.getSocket().getOutputStream());
+				OutputStream clienteOS = cliente.getSocket().getOutputStream();) {
 
 			while (entradaCliente.hasNextLine()) {
-
 				String comando = entradaCliente.nextLine();
-				
 				System.out.println(cliente.getSocket() + " - " + comando);
 
 				switch (comando.toLowerCase().split(" ")[0]) {
@@ -45,96 +48,47 @@ public class DistribuidorDeTarefas implements Runnable {
 				case "get": // get Aquivo B.txt > C:\Minha Pasta\
 					String[] parametros = comando.substring(4, comando.length()).split(" > ");
 					File returnFile = new File(cliente.getPathAtual() + parametros[0]);
-					
+
 					if (!returnFile.exists()) {
 						saidaCliente.println("Arquivo não encontrado:" + parametros[0]);
 						break;
 					}
-					
+
 					saidaCliente.println("file");
 					saidaCliente.println(parametros[0]);
 					saidaCliente.println(parametros[1]);
-					
-					byte [] fileByteArray  = new byte [(int)returnFile.length()];
-			        FileInputStream fis = new FileInputStream(returnFile);
-			        BufferedInputStream bis = new BufferedInputStream(fis);
-			        bis.read(fileByteArray,0,fileByteArray.length);
-					
-			        saidaCliente.write(fileByteArray,0,fileByteArray.length);
-					cliente.getSocket().getOutputStream().flush();										
+
+					byte[] fileByteArray = new byte[(int) returnFile.length()];
+					FileInputStream fis = new FileInputStream(returnFile);
+					BufferedInputStream bis = new BufferedInputStream(fis);
+					bis.read(fileByteArray, 0, fileByteArray.length);
+
+					saidaCliente.write(fileByteArray, 0, fileByteArray.length);
+					clienteOS.flush();
 					break;
 
 				case "rename": // rename Arquivo A.txt > Arquivo B.txt
-					String[] nomes = comando.substring(7, comando.length()).split(" > ");
-					File oldFile = new File(cliente.getPathAtual() + nomes[0]);
-					File newFile = new File(cliente.getPathAtual() + nomes[1]);
-					if (!newFile.exists()) {
-						if (oldFile.exists()) {
-							oldFile.renameTo(newFile);
-						} else {
-							saidaCliente.println("text");
-							saidaCliente.println("O arquivo " + nomes[0] + " não existe");
-						}
-					} else {
-						saidaCliente.println("text");
-						saidaCliente
-								.println("O nome de arquivo " + nomes[1] + " já existe, não foi possível renomea-lo!");
-					}
+					new Tarefas().renomearArquivo(saidaCliente, comando, cliente.getPathAtual());
 					break;
 
 				case "delete": // delete Arquivo A.txt
-					String path = removeUltimaBarra(cliente.getPathAtual());
-					String nome = removePrimeiraBarra(comando.substring(7, comando.length()));
-					boolean delete = new File(path + "/" + nome).delete();
-					if (delete) {
-						saidaCliente.println("text");
-						saidaCliente.println("Arquivo Removido com sucesso");
-					} else {
-						saidaCliente.println("text");
-						saidaCliente.println(
-								"Não foi possível excluir o Arquivo/Diretório especificado. Verifiquei o nome e se o mesmo existe.");
-					}
+					new Tarefas().deletar(saidaCliente, comando, cliente.getPathAtual());
 					break;
 
 				case "cd": // cd Minha Pasta
-					String pathNovo = comando.substring(3, comando.length());
-					String pathAtual = removeUltimaBarra(cliente.getPathAtual());
-					pathNovo = removePrimeiraBarra(pathNovo);
-					if (pathNovo.equals("..") && !isRaiz(cliente)) {
-						cliente.setPathAtual(pathAtual.substring(0, pathAtual.lastIndexOf("/")));
-					} else if (new File(pathAtual + "/" + pathNovo).exists() && !pathNovo.equals("..")) {
-						cliente.setPathAtual(pathAtual + "/" + pathNovo + "/");
-					} else {
-						saidaCliente.println("text");
-						saidaCliente.println("Diretório inválido!");
-					}
+					new Tarefas().navegar(saidaCliente, comando, cliente);
 					break;
 
 				case "ls": // ls
-					File[] listFiles = new File(cliente.getPathAtual()).listFiles();
-					for (int i = 0; i < listFiles.length; i++) {
-						if (listFiles[i].isDirectory()) {
-							saidaCliente.println("text");
-							saidaCliente.println(listFiles[i].getName() + " - Diretório");
-						} else {
-							saidaCliente.println("text");
-							saidaCliente.println(listFiles[i].getName() + " - Arquivo");
-						}
-					}
+					new Tarefas().listaArquivosDiretorioAtual(saidaCliente, cliente.getPathAtual());
 					break;
 
 				case "pwd": // pwd
-					saidaCliente.println("text");
-					saidaCliente.println(cliente.getPathAtual());
+					new Tarefas().exibirDiretorioAtual(saidaCliente, cliente.getPathAtual());
 					break;
 
 				case "mkdir": // mkdir Diretorio
-					File file = new File(cliente.getPathAtual() + comando.substring(6, comando.length()));
-					if (!file.exists()) {
-						file.mkdir();
-					} else {
-						saidaCliente.println("Diretório já existe!");
-					}
+					new Tarefas().cirarDiretorio(saidaCliente, comando, cliente.getPathAtual());
 					break;
 
 				default:
@@ -149,46 +103,6 @@ public class DistribuidorDeTarefas implements Runnable {
 			e.printStackTrace();
 		}
 
-	}
-
-	/**
-	 * Método auxiliar para remover a primeira barra
-	 * 
-	 * @param pathNovo
-	 *            - Novo Path para o qual o cliente irá
-	 * @return String - Novo Path sem a barra inicial
-	 */
-	private String removePrimeiraBarra(String pathNovo) {
-		if (pathNovo.charAt(0) == '/') {
-			return pathNovo.substring(1, pathNovo.length());
-		}
-		return pathNovo;
-	}
-
-	/**
-	 * Método auxiliar para remover a última barra do path
-	 * 
-	 * @param pathAtual
-	 *            - Caimnho onde o cliente está atualmente
-	 * @return String - Path sem a última barra
-	 */
-	private String removeUltimaBarra(String pathAtual) {
-		if (pathAtual.charAt(pathAtual.length() - 1) == '/') {
-			return pathAtual.substring(0, pathAtual.length() - 1);
-		}
-		return pathAtual;
-	}
-
-	/**
-	 * Método auxiliar que verifica se o Path atual do cliente é a raiz ou não
-	 * 
-	 * @param cliente
-	 *            - cliente contendo a informação do caminho que está atualmente
-	 * @return true - É a Raiz <br>
-	 *         false - Não é a Raiz
-	 */
-	private boolean isRaiz(Cliente cliente) {
-		return (cliente.getPathAtual().equalsIgnoreCase("./FTP/") || cliente.getPathAtual().equalsIgnoreCase("./FTP"));
 	}
 
 }
